Added functionality to dog, shelter, and home
The dog is now hidden initially and will be displayed when the game
starts. Added a victory card with a reset button. Added functions to
check if the dog is in the shelter or home, if it touches a
wall (review), and a winGame function. Added an event listener to pause
when the spacebar is pressed (review). Mouseover on the dog causes it
to follow the mouse cursor, and check if it is in the shelter, home, or
touching a wall every second. The reset button reloads the page.
The CSS styles now shrink and expand the game container. Added styles
for the victory card.