Added functionality to walls and fixed the dog
Added a failcard, dog starts at 1,1 , made more walls, dog resets
if it enters the home without an owner, checkDogTouchesWall function
checks each wall with a forEach loop (I wonder if this is the best way)
updated the space keydown event so it clears the failcard and resumes
the game. On mouseover, the dog now snaps to the mouse position.
On mousemove, the dog follows the mouse position, then checks if it
touches a wall, shelter, or home.
Added style to the failcard and adjusted some values for the dog.